Restructure 1delta tests
